chore(*): Refactor classes to retain `getState` as name. Use IGridStateInfo as interface name.
