+++ projects/igniteui-angular-elements/src/analyzer/elements.config.ts
@@ -440,8 +440,8 @@ export var registerConfig = [
     methods: [
       "applyState",
       "applyStateFromString",
-      "extractState",
-      "extractStateAsString",
+      "getState",
+      "getStateAsString",
     ],
   },
   {
