+++ projects/igniteui-angular-elements/src/index.html
@@ -179,7 +179,7 @@
 
     function saveState() {
         const stateComponent = document.getElementById('state');
-        const stringData = stateComponent.extractStateAsString();
+        const stringData = stateComponent.getStateAsString();
         window.localStorage.setItem(`grid1-state`, stringData);
     }
