+++ projects/igniteui-angular-elements/src/lib/state.component.ts
@@ -1,5 +1,4 @@
 import { Input, Component, EnvironmentInjector, Inject, Injector, ViewContainerRef } from '@angular/core';
-import {IgxGridStateDirective, IGridStateOptions, IGridState, GridFeatures, IColumnState, IGridStateCollection } from "../../../igniteui-angular/src/lib/grids/state.directive";
 import { IGX_GRID_SERVICE_BASE, GridServiceType, IPinningConfig}  from '../../../igniteui-angular/src/lib/grids/common/grid.interface';
 import { IgxGridComponent } from '../../../igniteui-angular/src/lib/grids/grid/grid.component';
 import { IFilteringExpressionsTree } from '../../../igniteui-angular/src/lib/data-operations/filtering-expressions-tree';
@@ -8,10 +7,11 @@ import { ISortingExpression } from '../../../igniteui-angular/src/lib/data-opera
 import { IGroupingState } from '../../../igniteui-angular/src/lib/data-operations/groupby-state.interface';
 import { GridSelectionRange } from '../../../igniteui-angular/src/lib/grids/common/types';
 import { IPivotConfiguration } from '../../../igniteui-angular/src/lib/grids/pivot-grid/pivot-grid.interface'
+import { GridFeatures, IColumnState, IGridStateCollection, IgxGridStateBaseDirective } from 'projects/igniteui-angular/src/lib/grids/state-base.directive';
 
 /* tsPlainInterface */
 /* marshalByValue */
-export interface IState {
+export interface IGridStateInfo {
     columns?: IColumnState[];
     filtering?: IFilteringExpressionsTree;
     advancedFiltering?: IFilteringExpressionsTree;
@@ -43,7 +43,7 @@ export interface IState {
     template: ``,
     standalone: true
 })
-export class IgxGridStateComponent extends IgxGridStateDirective {
+export class IgxGridStateComponent extends IgxGridStateBaseDirective {
 
     constructor(
         @Inject(IGX_GRID_SERVICE_BASE) private api: GridServiceType,
@@ -53,30 +53,8 @@ export class IgxGridStateComponent extends IgxGridStateDirective {
             super(api.grid, viewRef, envInjector, injector);
         }
 
-        @Input()
-        public override get options(): IGridStateOptions {
-           return this._options;
-        }
-
-        public override set options(value: IGridStateOptions) {
-            Object.assign(this._options, value);
-            if (!(this.grid instanceof IgxGridComponent)) {
-                delete this._options.groupBy;
-            } else {
-                delete this._options.rowIslands;
-            }
-        }
-
-
-    /** This method is marked as hidden, because it collides with the react components existing setState method.
-    /** Circumventing this by exposing a alternative applyState method. */
-    /** @hidden @internal */
-    public override setState(state: IGridState | string, features?: GridFeatures | GridFeatures[]) {
-        super.setState(state, features);
-    }
-
     /**
-     * Restores grid features' state based on the IGridState object passed as an argument.
+     * Restores grid features' state based on the IGridStateInfo object passed as an argument.
      *
      * @param IGridState object to restore state from.
      * @returns
@@ -88,58 +66,52 @@ export class IgxGridStateComponent extends IgxGridStateDirective {
      * this.state.applyState(gridState);
      * ```
      */
-    public applyState(state: IState , features: GridFeatures | GridFeatures[] = []) {
+    public applyState(state: IGridStateInfo , features: GridFeatures | GridFeatures[] = []) {
         if (features.length === 0) {
             features = null;
         }
-        super.setState(state, features);
-    }
-
-    /** @hidden @internal */
-    public override getState(serialize = true, features?: GridFeatures | GridFeatures[]): IGridState | string  {
-        return super.getState(serialize, features);
+        super.setStateInternal(state, features);
     }
 
     /**
      * Restores grid features' state based on the serialized IGridState object passed as an argument.
      *
-     * @param serialized IGridState object to restore state from.
      * @returns
      * ```html
      * <igx-grid [igxGridState]="options"></igx-grid>
      * ```
      * ```typescript
      * @ViewChild(IgxGridStateDirective, { static: true }) public state;
-     * this.state.setState(gridState);
+     * this.state.setStateFromString(gridState);
      * ```
      */
     public applyStateFromString(state: string, features: GridFeatures | GridFeatures[] = []) {
         if (features.length === 0) {
             features = null;
         }
-        super.setState(state, features);
+        super.setStateInternal(state, features);
     }
 
     /**
      * Gets the state of a feature or states of all grid features, unless a certain feature is disabled through the `options` property.
      *
      * @param `feature` string or array of strings determining the features to be added in the state. If skipped, all features are added.
-    * @returns Returns the non-serialized IGridState object.
+    * @returns Returns the non-serialized IGridStateInfo object.
      * ```html
      * <igx-grid [igxGridState]="options"></igx-grid>
      * ```
      * ```typescript
      * @ViewChild(IgxGridStateDirective, { static: true }) public state;
-     * let state = this.state.extractState(); // returns `IGridState` object
+     * let state = this.state.getState(); // returns `IGridStateInfo` object
      * ```
      */
-    public extractState(features: GridFeatures | GridFeatures[] = []): IState {
+    public getState(features: GridFeatures | GridFeatures[] = []): IGridStateInfo {
         /** Due to return type in getState being union type and having no support for union type in the translators
          * hiding getState in favor of a simpler extractState method that omits the serialize property and always returns just a string. */
         if (features.length === 0) {
             features = null;
         }
-        return super.getState(false, features) as IState;
+        return super.getStateInternal(false, features) as IGridStateInfo;
     }
 
     /**
@@ -152,14 +124,14 @@ export class IgxGridStateComponent extends IgxGridStateDirective {
      * ```
      * ```typescript
      * @ViewChild(IgxGridStateDirective, { static: true }) public state;
-     * let state = this.state.extractStateAsString(); // returns string
+     * let state = this.state.getStateAsString(); // returns string
      * ```
      */
-    public extractStateAsString(features: GridFeatures | GridFeatures[] = []): string {
+    public getStateAsString(features: GridFeatures | GridFeatures[] = []): string {
         if (features.length === 0) {
             features = null;
         }
-        return super.getState(true, features) as string;
+        return super.getStateInternal(true, features) as string;
     }
